services/metrics_collector/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
prometheus_client
influxdb-client